+++ tests/Feature/Service/Student/StudentListServiceTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Service;
+namespace Tests\Feature\Service\Student;
 
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
